requirements.txt
flask
gunicorn
requests
beautifulsoup4
pandas
ddgs
python-dotenv